garage

+++ garage/app/Http/Controllers/MechanicsController.php
@@ -42,7 +42,7 @@ class MechanicsController extends Controller
         $mechanics->name = $request->name;
         $mechanics->surname = $request->surname;
         $mechanics->save();
-        return redirect()->route('m_index');
+        return redirect()->route('m_index')->with('success_msg', 'Good job. We have new mechanic now.');
     }
 
     /**
@@ -83,7 +83,7 @@ class MechanicsController extends Controller
         $mechanics->name = $request->name;
         $mechanics->surname = $request->surname;
         $mechanics->save();
-        return redirect()->route('m_index');
+        return redirect()->route('m_index')->with('success_msg', 'Good job. Mechanic was updated.');
     }
 
     /**
@@ -94,6 +94,9 @@ class MechanicsController extends Controller
      */
     public function destroy(Mechanics $mechanics)
     {
+        if ($mechanics->getTrucks()->count()) {
+            return redirect()->back()->with('info_msg', 'Oh no, you can not delete this one.');
+        }
         $mechanics->delete();
         return redirect()->route('m_index');
     }
